jajaja

+++ database/migrations/2025_01_14_152524_remove_view_from_rooms_table.php
@@ -8,9 +8,11 @@ class RemoveViewFromRoomsTable extends Migration
 {
     public function up()
     {
-        Schema::table('rooms', function (Blueprint $table) {
-            $table->dropColumn('view');
-        });
+        if (Schema::hasColumn('rooms', 'view')) {
+            Schema::table('rooms', function (Blueprint $table) {
+                $table->dropColumn('view');
+            });
+        }
     }
 
     public function down()
@@ -19,4 +21,6 @@ class RemoveViewFromRoomsTable extends Migration
             $table->string('view')->nullable();
         });
     }
+
+
 }
